requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
beautifulsoup4==4.7.1
black==18.9b0
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dhooks==1.1.0
Flask==1.0.2
flup==1.0.3
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
multidict==4.5.2
peewee==3.8.2
python-dotenv==0.10.1
redis==3.2.0
requests==2.20.1
selenium==3.141.0
soupsieve==1.7.3
toml==0.10.0
urllib3==1.24.1
Werkzeug==0.14.1
yarl==1.3.0
